+++ Meetup.Contracts/Models/MeetupDto.cs
@@ -1,0 +1,13 @@
+﻿using System.ComponentModel.DataAnnotations;
+namespace MeetupAPI.Models;
+
+public class MeetupDto
+{
+    [Required]
+    [MinLength(3)]
+    public string Name { get; set; }
+    [Required]
+    public string Organizer { get; set; }
+    public DateTime Date { get; set; }
+    public bool IsPrivate { get; set; }
+}
