+++ Meetup.Contracts/Models/RegisterUserDto.cs
@@ -1,0 +1,10 @@
+﻿namespace MeetupAPI.Models;
+
+public class RegisterUserDto : UserLoginDto
+{
+    public string Nationality { get; set; }
+
+    public DateTime? DateOfBirth { get; set; }
+
+    public int RoleId { get; set; } = 1;
+}
