+++ Meetup.Contracts/Models/SortDirection.cs
@@ -1,0 +1,7 @@
+﻿namespace MeetupAPI.Models;
+
+public enum SortDirection
+{
+    ASC,
+    DESC
+}
