+++ Meetup.Contracts/Models/UpdateUserDto.cs
@@ -1,0 +1,14 @@
+﻿namespace MeetupAPI.Models;
+
+public class UpdateUserDto : UserLoginDto
+{
+    public string FirstName { get; set; }
+
+    public string LastName { get; set; }
+
+    public string Nationality { get; set; }
+
+    public DateTime? DateOfBirth { get; set; }
+
+    public int RoleId { get; set; } = 1;
+}
