+++ MeetupAPI.sln
@@ -9,7 +9,13 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "FillUpDatabaseWithNewMeetup
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Meetup.UnitTests", "Meetup.UnitTests\Meetup.UnitTests.csproj", "{C38EE72A-9EB6-4212-BE2B-B7D3DC857F6D}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Meetup.IntegrationTests", "Meetup.IntegrationTests\Meetup.IntegrationTests.csproj", "{D052807A-4BC1-4DF4-865F-60944D3FBB97}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Meetup.IntegrationTests", "Meetup.IntegrationTests\Meetup.IntegrationTests.csproj", "{D052807A-4BC1-4DF4-865F-60944D3FBB97}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Meetup.Sdk", "Meetup.Sdk\Meetup.Sdk.csproj", "{14CFD718-FDEF-4E6D-BFD0-9B1903930ED5}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Meetup.Contracts", "Meetup.Contracts\Meetup.Contracts.csproj", "{2BB72AF6-6BB0-46A7-89BF-DCCB93B7D90D}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Meetup.Sdk.Sample", "Meetup.Sdk.Sample\Meetup.Sdk.Sample.csproj", "{7C62B655-6350-4DD7-930A-A23B4069D088}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -33,6 +39,18 @@ Global
 		{D052807A-4BC1-4DF4-865F-60944D3FBB97}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{D052807A-4BC1-4DF4-865F-60944D3FBB97}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{D052807A-4BC1-4DF4-865F-60944D3FBB97}.Release|Any CPU.Build.0 = Release|Any CPU
+		{14CFD718-FDEF-4E6D-BFD0-9B1903930ED5}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{14CFD718-FDEF-4E6D-BFD0-9B1903930ED5}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{14CFD718-FDEF-4E6D-BFD0-9B1903930ED5}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{14CFD718-FDEF-4E6D-BFD0-9B1903930ED5}.Release|Any CPU.Build.0 = Release|Any CPU
+		{2BB72AF6-6BB0-46A7-89BF-DCCB93B7D90D}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{2BB72AF6-6BB0-46A7-89BF-DCCB93B7D90D}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{2BB72AF6-6BB0-46A7-89BF-DCCB93B7D90D}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{2BB72AF6-6BB0-46A7-89BF-DCCB93B7D90D}.Release|Any CPU.Build.0 = Release|Any CPU
+		{7C62B655-6350-4DD7-930A-A23B4069D088}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{7C62B655-6350-4DD7-930A-A23B4069D088}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{7C62B655-6350-4DD7-930A-A23B4069D088}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{7C62B655-6350-4DD7-930A-A23B4069D088}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
