+++ MeetupAPI/Authorization/MeetupResourceOperationHandler.cs
@@ -3,25 +3,24 @@ using Microsoft.AspNetCore.Authorization;
 using System.Security.Claims;
 using System.Threading.Tasks;
 
-namespace MeetupAPI.Authorization
+namespace MeetupAPI.Authorization;
+
+public class MeetupResourceOperationHandler : AuthorizationHandler<ResourceOperationRequirement, Meetup>
 {
-    public class MeetupResourceOperationHandler : AuthorizationHandler<ResourceOperationRequirement, Meetup>
+    protected override Task HandleRequirementAsync(AuthorizationHandlerContext context, ResourceOperationRequirement requirement, Meetup resource)
     {
-        protected override Task HandleRequirementAsync(AuthorizationHandlerContext context, ResourceOperationRequirement requirement, Meetup resource)
+        if (requirement.OperationType == OperationType.Create || requirement.OperationType == OperationType.Read)
         {
-            if (requirement.OperationType == OperationType.Create || requirement.OperationType == OperationType.Read)
-            {
-                context.Succeed(requirement);
-            }
-
-            var userId = context.User.FindFirst(c => c.Type == ClaimTypes.NameIdentifier).Value;
-
-            if (resource.CreatedById == int.Parse(userId))
-            {
-                context.Succeed(requirement);
-            }
-
-            return Task.CompletedTask;
+            context.Succeed(requirement);
         }
+
+        var userId = context.User.FindFirst(c => c.Type == ClaimTypes.NameIdentifier).Value;
+
+        if (resource.CreatedById == int.Parse(userId))
+        {
+            context.Succeed(requirement);
+        }
+
+        return Task.CompletedTask;
     }
 }
