+++ MeetupAPI/Authorization/MinimumAgeRequirement.cs
@@ -1,14 +1,13 @@
 ﻿using Microsoft.AspNetCore.Authorization;
 
-namespace MeetupAPI.Authorization
-{
-    public class MinimumAgeRequirement : IAuthorizationRequirement
-    {
-        public int MinimumAge { get; }
+namespace MeetupAPI.Authorization;
 
-        public MinimumAgeRequirement(int minimumAge)
-        {
-            MinimumAge = minimumAge;
-        }
+public class MinimumAgeRequirement : IAuthorizationRequirement
+{
+    public int MinimumAge { get; }
+
+    public MinimumAgeRequirement(int minimumAge)
+    {
+        MinimumAge = minimumAge;
     }
 }
