+++ MeetupAPI/Authorization/ResourceOperationRequirement.cs
@@ -1,22 +1,21 @@
 ﻿using Microsoft.AspNetCore.Authorization;
 
-namespace MeetupAPI.Authorization
+namespace MeetupAPI.Authorization;
+
+public enum OperationType
 {
-    public enum OperationType
+    Create,
+    Read,
+    Update,
+    Delete
+}
+
+public class ResourceOperationRequirement : IAuthorizationRequirement
+{
+    public ResourceOperationRequirement(OperationType operationType)
     {
-        Create,
-        Read,
-        Update,
-        Delete
+        OperationType = operationType;
     }
 
-    public class ResourceOperationRequirement : IAuthorizationRequirement
-    {
-        public ResourceOperationRequirement(OperationType operationType)
-        {
-            OperationType = operationType;
-        }
-
-        public OperationType OperationType { get; }
-    }
+    public OperationType OperationType { get; }
 }
