+++ MeetupAPI/Controllers/ConfigController.cs
@@ -2,30 +2,29 @@
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.Extensions.Configuration;
 
-namespace MeetupAPI.Controllers
+namespace MeetupAPI.Controllers;
+
+[Route("config")]
+public class ConfigController : ControllerBase
 {
-    [Route("config")]
-    public class ConfigController : ControllerBase
+    private readonly IConfiguration _configuration;
+    public ConfigController(IConfiguration configuration)
     {
-        private readonly IConfiguration _configuration;
-        public ConfigController(IConfiguration configuration)
-        {
-            _configuration = configuration;
-        }
-
-        [HttpOptions("reload")]
-        public ActionResult Reload()
-        {
-            try
-            {
-                ((IConfigurationRoot)_configuration).Reload();
-                return Ok();
-            }
-            catch (Exception)
-            {
-                return StatusCode(500);
-            }
-        }
-
+        _configuration = configuration;
     }
+
+    [HttpOptions("reload")]
+    public ActionResult Reload()
+    {
+        try
+        {
+            ((IConfigurationRoot)_configuration).Reload();
+            return Ok();
+        }
+        catch (Exception)
+        {
+            return StatusCode(500);
+        }
+    }
+
 }
