+++ MeetupAPI/Controllers/MeetupController.cs
@@ -12,169 +12,168 @@ using Microsoft.EntityFrameworkCore;
 using System.Linq.Expressions;
 using System;
 
-namespace MeetupAPI.Controllers
+namespace MeetupAPI.Controllers;
+
+[Route("api/meetup")]
+[Authorize]
+[ServiceFilter(typeof(TimeTrackFilter))]
+public class MeetupController : ControllerBase
 {
-    [Route("api/meetup")]
-    [Authorize]
-    [ServiceFilter(typeof(TimeTrackFilter))]
-    public class MeetupController : ControllerBase
+    private readonly MeetupContext _meetupContext;
+    private readonly IMapper _mapper;
+    private readonly IAuthorizationService _authorizationService;
+
+    public MeetupController(MeetupContext meetupContext, IMapper mapper, IAuthorizationService authorizationService)
     {
-        private readonly MeetupContext _meetupContext;
-        private readonly IMapper _mapper;
-        private readonly IAuthorizationService _authorizationService;
+        _meetupContext = meetupContext;
+        _mapper = mapper;
+        _authorizationService = authorizationService;
+    }
 
-        public MeetupController(MeetupContext meetupContext, IMapper mapper, IAuthorizationService authorizationService)
+    [HttpGet]
+    //[NationalityFilter("German,Russian")]
+    [AllowAnonymous]
+    public ActionResult<PagedResult<MeetupDetailsDto>> GetAll([FromQuery]MeetupQuery query)
+    {
+        if (!ModelState.IsValid)
         {
-            _meetupContext = meetupContext;
-            _mapper = mapper;
-            _authorizationService = authorizationService;
+            return BadRequest(ModelState);
+        }
+        
+        var baseQuery = _meetupContext.Meetups
+            .Include(m => m.Location)
+            .Where(m => query.SearchPhrase == null ||
+                        m.Organizer.ToLower().Contains(query.SearchPhrase.ToLower()) || 
+                        m.Name.Contains(query.SearchPhrase.ToLower()));
+
+        if (!string.IsNullOrEmpty(query.SortBy))
+        {
+            var propertySelectors = new Dictionary<string, Expression<Func<Meetup, object>>>()
+            {
+                { nameof(Meetup.Name), meetup => meetup.Name },
+                { nameof(Meetup.Date), meetup => meetup.Date },
+                { nameof(Meetup.Organizer), meetup => meetup.Organizer }
+            };
+
+            var propertySelector = propertySelectors[query.SortBy];
+
+            baseQuery = query.SortDirection == SortDirection.ASC
+                ? baseQuery.OrderBy(m => m.Date)
+                : baseQuery.OrderByDescending(m => m.Name);
+        }
+        
+        var meetups = baseQuery
+            .Skip(query.PageSize * (query.PageNumber - 1))
+            .Take(query.PageSize)
+            .ToList();
+
+        var totalCount = baseQuery.Count();
+        
+        var meetupDtos = _mapper.Map<List<MeetupDetailsDto>>(meetups);
+
+        var result = new PagedResult<MeetupDetailsDto>(meetupDtos, totalCount, query.PageNumber, query.PageSize);
+        
+        return Ok(result);
+    }
+
+    [HttpGet("{name}")]
+    //[NationalityFilter("English")]
+    // Authorisation policy that checks if the user has a claim nationality
+    //[Authorize(Policy = "HasNationality")]
+    //[Authorize(Policy = "AtLeast18")]
+    public ActionResult<MeetupDetailsDto> Get(string name)
+    {
+        var meetup = _meetupContext.Meetups
+            .Include(m => m.Location)
+            .Include(m => m.Lectures)
+            .FirstOrDefault(m => m.Name.Replace(" ", "-").ToLower() == name.ToLower());
+
+        if (meetup == null)
+        {
+            return NotFound();
         }
 
-        [HttpGet]
-        //[NationalityFilter("German,Russian")]
-        [AllowAnonymous]
-        public ActionResult<PagedResult<MeetupDetailsDto>> GetAll([FromQuery]MeetupQuery query)
+        var meetupDto = _mapper.Map<MeetupDetailsDto>(meetup);
+        return Ok(meetupDto);
+    }
+
+    [HttpPost]
+    [Authorize(Roles = "Admin,Moderator")]
+    public ActionResult Post([FromBody]MeetupDto model)
+    {
+        if (!ModelState.IsValid)
         {
-            if (!ModelState.IsValid)
-            {
-                return BadRequest(ModelState);
-            }
-            
-            var baseQuery = _meetupContext.Meetups
-                .Include(m => m.Location)
-                .Where(m => query.SearchPhrase == null ||
-                            m.Organizer.ToLower().Contains(query.SearchPhrase.ToLower()) || 
-                            m.Name.Contains(query.SearchPhrase.ToLower()));
-
-            if (!string.IsNullOrEmpty(query.SortBy))
-            {
-                var propertySelectors = new Dictionary<string, Expression<Func<Meetup, object>>>()
-                {
-                    { nameof(Meetup.Name), meetup => meetup.Name },
-                    { nameof(Meetup.Date), meetup => meetup.Date },
-                    { nameof(Meetup.Organizer), meetup => meetup.Organizer }
-                };
-
-                var propertySelector = propertySelectors[query.SortBy];
-
-                baseQuery = query.SortDirection == SortDirection.ASC
-                    ? baseQuery.OrderBy(m => m.Date)
-                    : baseQuery.OrderByDescending(m => m.Name);
-            }
-            
-            var meetups = baseQuery
-                .Skip(query.PageSize * (query.PageNumber - 1))
-                .Take(query.PageSize)
-                .ToList();
-
-            var totalCount = baseQuery.Count();
-            
-            var meetupDtos = _mapper.Map<List<MeetupDetailsDto>>(meetups);
-
-            var result = new PagedResult<MeetupDetailsDto>(meetupDtos, totalCount, query.PageNumber, query.PageSize);
-            
-            return Ok(result);
+            return BadRequest(ModelState);
         }
 
-        [HttpGet("{name}")]
-        //[NationalityFilter("English")]
-        // Authorisation policy that checks if the user has a claim nationality
-        //[Authorize(Policy = "HasNationality")]
-        //[Authorize(Policy = "AtLeast18")]
-        public ActionResult<MeetupDetailsDto> Get(string name)
+        var meetup = _mapper.Map<Meetup>(model);
+
+        var userId = User.FindFirst(c => c.Type == ClaimTypes.NameIdentifier).Value;
+
+        meetup.CreatedById = int.Parse(userId);
+
+        _meetupContext.Meetups.Add(meetup);
+        _meetupContext.SaveChanges();
+
+        var key = meetup.Name.Replace(" ", "-").ToLower();
+        return Created("api/meetup/" + key, null);
+    }
+
+    [HttpPut("{name}")]
+    public ActionResult Put(string name, [FromBody] MeetupDto model)
+    {
+        var meetup = _meetupContext.Meetups
+            .FirstOrDefault(m => m.Name.Replace(" ", "-").ToLower() == name.ToLower());
+
+        if (meetup == null)
         {
-            var meetup = _meetupContext.Meetups
-                .Include(m => m.Location)
-                .Include(m => m.Lectures)
-                .FirstOrDefault(m => m.Name.Replace(" ", "-").ToLower() == name.ToLower());
-
-            if (meetup == null)
-            {
-                return NotFound();
-            }
-
-            var meetupDto = _mapper.Map<MeetupDetailsDto>(meetup);
-            return Ok(meetupDto);
+            return NotFound();
         }
 
-        [HttpPost]
-        [Authorize(Roles = "Admin,Moderator")]
-        public ActionResult Post([FromBody]MeetupDto model)
+        var authorizationResult = _authorizationService.AuthorizeAsync(User, meetup, new ResourceOperationRequirement(OperationType.Update)).Result;
+
+        if (!authorizationResult.Succeeded)
         {
-            if (!ModelState.IsValid)
-            {
-                return BadRequest(ModelState);
-            }
-
-            var meetup = _mapper.Map<Meetup>(model);
-
-            var userId = User.FindFirst(c => c.Type == ClaimTypes.NameIdentifier).Value;
-
-            meetup.CreatedById = int.Parse(userId);
-
-            _meetupContext.Meetups.Add(meetup);
-            _meetupContext.SaveChanges();
-
-            var key = meetup.Name.Replace(" ", "-").ToLower();
-            return Created("api/meetup/" + key, null);
+            return Forbid();
+        }
+        
+        if (!ModelState.IsValid)
+        {
+            return BadRequest(ModelState);
         }
 
-        [HttpPut("{name}")]
-        public ActionResult Put(string name, [FromBody] MeetupDto model)
+        meetup.Name = model.Name;
+        meetup.Organizer = model.Organizer;
+        meetup.Date = model.Date;
+        meetup.IsPrivate = model.IsPrivate;
+
+        _meetupContext.SaveChanges();
+
+        return NoContent();
+    }
+
+    [HttpDelete("{name}")]
+    public ActionResult Delete(string name)
+    {
+        var meetup = _meetupContext.Meetups
+            .Include(m => m.Location)
+            .FirstOrDefault(m => m.Name.Replace(" ", "-").ToLower() == name.ToLower());
+
+        if (meetup == null)
         {
-            var meetup = _meetupContext.Meetups
-                .FirstOrDefault(m => m.Name.Replace(" ", "-").ToLower() == name.ToLower());
-
-            if (meetup == null)
-            {
-                return NotFound();
-            }
-
-            var authorizationResult = _authorizationService.AuthorizeAsync(User, meetup, new ResourceOperationRequirement(OperationType.Update)).Result;
-
-            if (!authorizationResult.Succeeded)
-            {
-                return Forbid();
-            }
-            
-            if (!ModelState.IsValid)
-            {
-                return BadRequest(ModelState);
-            }
-
-            meetup.Name = model.Name;
-            meetup.Organizer = model.Organizer;
-            meetup.Date = model.Date;
-            meetup.IsPrivate = model.IsPrivate;
-
-            _meetupContext.SaveChanges();
-
-            return NoContent();
+            return NotFound();
         }
 
-        [HttpDelete("{name}")]
-        public ActionResult Delete(string name)
+        var authorizationResult = _authorizationService.AuthorizeAsync(User, meetup, new ResourceOperationRequirement(OperationType.Delete)).Result;
+
+        if (!authorizationResult.Succeeded)
         {
-            var meetup = _meetupContext.Meetups
-                .Include(m => m.Location)
-                .FirstOrDefault(m => m.Name.Replace(" ", "-").ToLower() == name.ToLower());
-
-            if (meetup == null)
-            {
-                return NotFound();
-            }
-
-            var authorizationResult = _authorizationService.AuthorizeAsync(User, meetup, new ResourceOperationRequirement(OperationType.Delete)).Result;
-
-            if (!authorizationResult.Succeeded)
-            {
-                return Forbid();
-            }
-
-            _meetupContext.Remove(meetup);
-            _meetupContext.SaveChanges();
-
-            return NoContent();
+            return Forbid();
         }
+
+        _meetupContext.Remove(meetup);
+        _meetupContext.SaveChanges();
+
+        return NoContent();
     }
 }
