+++ MeetupAPI/Entities/Lecture.cs
@@ -1,19 +1,12 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
+﻿namespace MeetupAPI.Entities;
 
-namespace MeetupAPI.Entities
+public class Lecture
 {
-    public class Lecture
-    {
-        public int Id { get; set; }
-        public string Author { get; set; }
-        public string Topic { get; set; }
-        public string Description { get; set; }
+    public int Id { get; set; }
+    public string Author { get; set; }
+    public string Topic { get; set; }
+    public string Description { get; set; }
 
-        public virtual Meetup Meetup { get; set; }
-        public int MeetupId { get; set; }
-    }
+    public virtual Meetup Meetup { get; set; }
+    public int MeetupId { get; set; }
 }
