+++ MeetupAPI/Entities/Meetup.cs
@@ -1,18 +1,17 @@
 ﻿using System;
 using System.Collections.Generic;
 
-namespace MeetupAPI.Entities
+namespace MeetupAPI.Entities;
+
+public class Meetup
 {
-    public class Meetup
-    {
-        public int Id { get; set; }
-        public string Name { get; set; }
-        public string Organizer { get; set; }
-        public DateTime Date { get; set; }
-        public bool IsPrivate { get; set; }
-        public virtual Location Location { get; set; }
-        public virtual List<Lecture> Lectures { get; set; }
-        public int? CreatedById { get; set; }
-        public User CreatedBy { get; set; }
-    }
+    public int Id { get; set; }
+    public string Name { get; set; }
+    public string Organizer { get; set; }
+    public DateTime Date { get; set; }
+    public bool IsPrivate { get; set; }
+    public virtual Location Location { get; set; }
+    public virtual List<Lecture> Lectures { get; set; }
+    public int? CreatedById { get; set; }
+    public User CreatedBy { get; set; }
 }
