+++ MeetupAPI/Entities/Role.cs
@@ -1,8 +1,7 @@
-﻿namespace MeetupAPI.Entities
+﻿namespace MeetupAPI.Entities;
+
+public class Role
 {
-    public class Role
-    {
-        public int Id { get; set; }
-        public string RoleName { get; set; }
-    }
+    public int Id { get; set; }
+    public string RoleName { get; set; }
 }
