+++ MeetupAPI/Entities/User.cs
@@ -1,17 +1,16 @@
 ﻿using System;
 
-namespace MeetupAPI.Entities
+namespace MeetupAPI.Entities;
+
+public class User
 {
-    public class User
-    {
-        public int Id { get; set; }
-        public string Email { get; set; }
-        public string FirstName { get; set; }
-        public string LastName { get; set; }
-        public string Nationality { get; set; }
-        public DateTime? DateOfBirth { get; set; }
-        public string PasswordHash { get; set; }
-        public int RoleId { get; set; }
-        public Role Role { get; set; } 
-    }
+    public int Id { get; set; }
+    public string Email { get; set; }
+    public string FirstName { get; set; }
+    public string LastName { get; set; }
+    public string Nationality { get; set; }
+    public DateTime? DateOfBirth { get; set; }
+    public string PasswordHash { get; set; }
+    public int RoleId { get; set; }
+    public Role Role { get; set; } 
 }
