+++ MeetupAPI/Filters/ExceptionFilter.cs
@@ -2,26 +2,25 @@
 using Microsoft.AspNetCore.Mvc.Filters;
 using Microsoft.Extensions.Logging;
 
-namespace MeetupAPI.Filters
+namespace MeetupAPI.Filters;
+
+public class ExceptionFilter : IExceptionFilter
 {
-    public class ExceptionFilter : IExceptionFilter
+    private readonly ILogger<ExceptionFilter> _logger;
+
+    public ExceptionFilter(ILogger<ExceptionFilter> logger)
     {
-        private readonly ILogger<ExceptionFilter> _logger;
-
-        public ExceptionFilter(ILogger<ExceptionFilter> logger)
-        {
-            _logger = logger;
-        }
+        _logger = logger;
+    }
 #if DEBUG
-        public void OnException(ExceptionContext context)
-        {
-            _logger.LogCritical($"Caught in ExceptionFilter {context.Exception.Message}", context.Exception);
+    public void OnException(ExceptionContext context)
+    {
+        _logger.LogCritical($"Caught in ExceptionFilter {context.Exception.Message}", context.Exception);
 
-            var result = new JsonResult("Something went wrong");
-            result.StatusCode = 500;
+        var result = new JsonResult("Something went wrong");
+        result.StatusCode = 500;
 
-            context.Result = result;
-        }
+        context.Result = result;
     }
 }
 #endif
