+++ MeetupAPI/Identity/JwtOptions.cs
@@ -1,9 +1,8 @@
-﻿namespace MeetupAPI.Identity
+﻿namespace MeetupAPI.Identity;
+
+public class JwtOptions
 {
-    public class JwtOptions
-    {
-        public string JwtKey { get; set; }
-        public string JwtIssuer { get; set; }
-        public int JwtExpireDays { get; set; }
-    }
+    public string JwtKey { get; set; }
+    public string JwtIssuer { get; set; }
+    public int JwtExpireDays { get; set; }
 }
