+++ MeetupAPI/MeetupProfile.cs
@@ -1,27 +1,21 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-using AutoMapper;
+﻿using AutoMapper;
 using MeetupAPI.Entities;
 using MeetupAPI.Models;
 
-namespace MeetupAPI
+namespace MeetupAPI;
+
+public class MeetupProfile : Profile
 {
-    public class MeetupProfile : Profile
+    public MeetupProfile()
     {
-        public MeetupProfile()
-        {
-            CreateMap<Meetup, MeetupDetailsDto>()
-                .ForMember(m => m.City, map => map.MapFrom(meetup => meetup.Location.City))
-                .ForMember(m => m.PostCode, map => map.MapFrom(meetup => meetup.Location.PostCode))
-                .ForMember(m => m.Street, map => map.MapFrom(meetup => meetup.Location.Street));
+        CreateMap<Meetup, MeetupDetailsDto>()
+            .ForMember(m => m.City, map => map.MapFrom(meetup => meetup.Location.City))
+            .ForMember(m => m.PostCode, map => map.MapFrom(meetup => meetup.Location.PostCode))
+            .ForMember(m => m.Street, map => map.MapFrom(meetup => meetup.Location.Street));
 
-            CreateMap<MeetupDto, Meetup>();
+        CreateMap<MeetupDto, Meetup>();
 
-            CreateMap<LectureDto, Lecture>()
-                .ReverseMap();
-        }
+        CreateMap<LectureDto, Lecture>()
+            .ReverseMap();
     }
 }
