--- MeetupAPI/Models/MeetupDto.cs
@@ -1,15 +1,0 @@
-﻿using System;
-using System.ComponentModel.DataAnnotations;
-namespace MeetupAPI.Models
-{
-    public class MeetupDto
-    {
-        [Required]
-        [MinLength(3)]
-        public string Name { get; set; }
-        [Required]
-        public string Organizer { get; set; }
-        public DateTime Date { get; set; }
-        public bool IsPrivate { get; set; }
-    }
-}
