--- MeetupAPI/Models/RegisterUserDto.cs
@@ -1,13 +1,0 @@
-﻿using System;
-
-namespace MeetupAPI.Models
-{
-    public class RegisterUserDto : UserLoginDto
-    {
-        public string Nationality { get; set; }
-
-        public DateTime? DateOfBirth { get; set; }
-
-        public int RoleId { get; set; } = 1;
-    }
-}
