--- MeetupAPI/Models/SortDirection.cs
@@ -1,8 +1,0 @@
-﻿namespace MeetupAPI.Models
-{
-    public enum SortDirection
-    {
-        ASC,
-        DESC
-    }
-}
