--- MeetupAPI/Models/UserLoginDto.cs
@@ -1,9 +1,0 @@
-﻿namespace MeetupAPI.Models
-{
-    public class UserLoginDto
-    {
-        public string Email { get; set; }
-        public string Password { get; set; }
-        public string ConfirmPassword { get; set; }
-    }
-}
